Wallet fix

+++ penny_wise/api/models.py
@@ -93,7 +93,7 @@ class Transaction(models.Model):
         on_delete=models.CASCADE,
         related_name='transactions'
     )
-    amount = models.IntegerField()
+    amount = models.DecimalField(max_digits=20, decimal_places=2)
     comment = models.TextField(blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
